refactor(website): re-organize messy frontend imports
