+++ apps/website/src/app/dashboard/[id]/modmail/panels/_components/CategoryCheckboxList.tsx
